+++ packages/forma-36-codemod/transforms/__testfixtures__/color-tokens-to-new-tokens.input.js
@@ -1,0 +1,34 @@
+import tokens from '@contentful/forma-36-tokens';
+
+const colorElementMid = '#eeeeee';
+
+export const styles = {
+  filterPill: {
+    color: colorElementMid,
+    backgroundColor: tokens.colorElementMid,
+  },
+  filterPillLabel: {
+    backgroundColor: tokens.colorElementLight,
+    color: tokens.colorTextMid,
+  },
+  filterSelectContainerActive: {
+    backgroundColor: tokens.colorBlueMid,
+    color: tokens.colorWhite,
+  },
+  filterSelect: {
+    color: tokens.colorTextMid,
+    backgroundColor: tokens.colorElementMid,
+    '&:hover': {
+      cursor: 'pointer',
+      backgroundColor: tokens.colorElementDark,
+    },
+  },
+  filterSelectActive: {
+    backgroundColor: tokens.colorBlueMid,
+    color: tokens.colorWhite,
+    '&:hover': {
+      backgroundColor: tokens.colorBlueBase,
+      color: tokens.colorWhite,
+    },
+  },
+};