+++ packages/forma-36-react-components/src/components/Icon/constants.ts
@@ -110,6 +110,7 @@ export const iconName = {
   PlusCircleTrimmed: 'PlusCircleTrimmed',
   PlusTrimmed: 'PlusTrimmed',
   Preview: 'Preview',
+  Puzzle: 'Puzzle',
   Quote: 'Quote',
   QuoteTrimmed: 'QuoteTrimmed',
   Receipt: 'Receipt',