+++ packages/forma-36-tokens/src/tokens/colors/colors-blue.js
@@ -1,9 +1,7 @@
-const colorsBlue = {
-  'color-blue-dark': '#174E8C',
-  'color-blue-base': '#2D64B3',
-  'color-blue-mid': '#2E75D4',
-  'color-blue-light': '#84B9F5',
-  'color-blue-lightest': '#EDF4FC',
-};
+const scaleBlue = require('./scale-blue');
+const colorsBlue = require('./colors-blue-deprecated');
 
-module.exports = colorsBlue;
+module.exports = {
+  ...colorsBlue,
+  ...scaleBlue,
+};