+++ packages/forma-36-tokens/src/tokens/colors/colors-contrast.js
@@ -1,7 +1,3 @@
-const colorsContrast = {
-  'color-contrast-dark': '#0C141C',
-  'color-contrast-mid': '#192532',
-  'color-contrast-light': '#263545',
-};
+const colorsContrast = require('./colors-contrast-deprecated');
 
 module.exports = colorsContrast;