+++ packages/forma-36-tokens/src/tokens/colors/colors-ice-deprecated.js
@@ -1,0 +1,8 @@
+const colorsBlue = require('./colors-blue');
+
+const colorsIce = {
+  'color-ice-dark': colorsBlue['blue-200'],
+  'color-ice-mid': colorsBlue['blue-100'],
+};
+
+module.exports = colorsIce;