+++ packages/forma-36-tokens/src/tokens/colors/colors-mint-deprecated.js
@@ -1,0 +1,8 @@
+const colorsGreen = require('./colors-green');
+
+const colorsMint = {
+  'color-mint-dark': colorsGreen['green-200'],
+  'color-mint-mid': colorsGreen['green-100'],
+};
+
+module.exports = colorsMint;