+++ packages/forma-36-tokens/src/tokens/colors/colors-orange.js
@@ -1,9 +1,7 @@
-const colorsOrange = {
-  'color-orange-dark': '#A85701',
-  'color-orange-base': '#DE8907',
-  'color-orange-mid': '#F79B0C',
-  'color-orange-light': '#FFD7A2',
-  'color-orange-lightest': '#FFE8C7',
-};
+const scaleOrange = require('./scale-orange');
+const colorsOrange = require('./colors-orange-deprecated');
 
-module.exports = colorsOrange;
+module.exports = {
+  ...colorsOrange,
+  ...scaleOrange,
+};