+++ packages/forma-36-tokens/src/tokens/colors/colors-peach-deprecated.js
@@ -1,0 +1,8 @@
+const colorsOrange = require('./colors-orange');
+
+const colorsPeach = {
+  'color-peach-dark': colorsOrange['orange-200'],
+  'color-peach-mid': colorsOrange['orange-100'],
+};
+
+module.exports = colorsPeach;