+++ packages/forma-36-tokens/src/tokens/colors/colors-purple.js
@@ -1,9 +1,7 @@
-const colorsPurple = {
-  'color-purple-dark': '#6426A2',
-  'color-purple-base': '#7439B0',
-  'color-purple-mid': '#8B4DC9',
-  'color-purple-light': '#D3ADF6',
-  'color-purple-lightest': '#F1E2FF',
-};
+const scalePurple = require('./scale-purple');
+const colorsPurple = require('./colors-purple-deprecated');
 
-module.exports = colorsPurple;
+module.exports = {
+  ...colorsPurple,
+  ...scalePurple,
+};