+++ packages/forma-36-tokens/src/tokens/colors/colors-red.js
@@ -1,9 +1,7 @@
-const colorsRed = {
-  'color-red-dark': '#7C262F',
-  'color-red-base': '#A82D3E',
-  'color-red-mid': '#BF3045',
-  'color-red-light': '#ECA7A7',
-  'color-red-lightest': '#FCE9E8',
-};
+const scaleRed = require('./scale-red');
+const colorsRed = require('./colors-red-deprecated');
 
-module.exports = colorsRed;
+module.exports = {
+  ...colorsRed,
+  ...scaleRed,
+};