+++ packages/forma-36-tokens/src/tokens/colors/scale-blue.js
@@ -1,0 +1,13 @@
+const scaleBlue = {
+  'blue-100': '#E8F5FF',
+  'blue-200': '#CEECFF',
+  'blue-300': '#98CBFF',
+  'blue-400': '#40A0FF',
+  'blue-500': '#036FE3',
+  'blue-600': '#0059C8',
+  'blue-700': '#0041AB',
+  'blue-800': '#003298',
+  'blue-900': '#002A8E',
+};
+
+module.exports = scaleBlue;