+++ packages/forma-36-website/gatsby-config.js
@@ -1,20 +1,25 @@
 const path = require('path');
 
+require('dotenv').config({
+  path: `.env.${process.env.NODE_ENV}`,
+});
+
 module.exports = {
   siteMetadata: {
     title: 'Forma 36 - The Contentful Design System',
-    promoText: '',
-    promoLink: '',
-    promoLinkText: '',
-    promoTagText: '',
+    promoText: `Updated color system with new tokens.`,
+    promoLink:
+      'https://github.com/contentful/forma-36/blob/master/packages/forma-36-tokens/MIGRATION.md',
+    promoLinkText: 'Migration notes',
+    promoTagText: 'New colors',
     menuLinks: [
       {
         name: 'Foundation',
         link: '',
         menuLinks: [
           {
-            name: 'Color',
-            link: '/foundation/color/',
+            name: 'Color system',
+            link: '/foundation/color-system/',
           },
           {
             name: 'Typography',